refactor(dashboard): extract logs/execution/governance domains (A1 phase 2)
- dashboard-logs.ts: fetchLogs/providerLogs/dashboardConfig/parseContextThresholds
- dashboard-execution.ts: execution trust/tool quality/perf/memory (+inline types)
- dashboard-governance.ts: governance approvals/HITL/schedule/case fetchers
- dashboard.ts: facade re-export only; consumers unchanged
- ripple import cleanup: normalizePendingConfirmation, withRetries,
  asKeeperApprovalRiskLevel, board normalize imports, ../types governance types
- typecheck green (tsc --noEmit)

dashboard.ts 3713 -> ~3300 lines after phase 2 (3 domains).
